move bugsnag
